requirements.txt
# AI Job Hunt - Requirements

# HTTP and web scraping
requests>=2.28.0
beautifulsoup4>=4.12.0
selenium>=4.9.0
playwright>=1.32.0
httpx>=0.24.0
webdriver-manager>=4.0.0

# Document processing
pdfplumber>=0.9.0
python-docx>=0.8.11
PyPDF2>=3.0.0
docx2txt>=0.8
nltk>=3.8.1
openpyxl>=3.1.2

# Data processing
pandas>=2.0.0
numpy>=1.24.0
pillow>=10.0.0

# Backend and API
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=2.0.0
python-multipart>=0.0.6
jinja2>=3.1.2
aiofiles>=23.1.0
email-validator>=2.0.0

# Database
sqlalchemy>=2.0.0
alembic>=1.11.0
aiosqlite>=0.19.0

# LLM and AI
openai>=1.0.0
langchain>=0.0.267

# UI
streamlit>=1.22.0
plotly>=5.14.0

# Utilities
schedule>=1.2.0
python-dotenv>=1.0.0
tqdm>=4.65.0
colorama>=0.4.6
rich>=13.3.5

# Networking
aiohttp>=3.8.5

# Testing
pytest>=7.3.1
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Development tools
black>=23.3.0
isort>=5.12.0
mypy>=1.2.0
flake8>=6.0.0
